GA-44 split out the authentication views
