+++ statistics/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from . import views
+from views import views, auth
 
 #The URLs are quite important to the GA Stats system, as there's a custom API and system for accessing various forms
 #Sessions, Debates, Point Submit and Vote Submit pages are accessed using the id's of the session/committee.
@@ -8,8 +8,8 @@ from . import views
 
 urlpatterns = [
     url(r'^$', views.home, name='home'),
-    url(r'^login/$', views.ga_login, name='login'),
-    url(r'^logout/$', views.ga_logout, name='logout'),
+    url(r'^login/$', auth.ga_login, name='login'),
+    url(r'^logout/$', auth.ga_logout, name='logout'),
     url(r'^create_session/$', views.create_session, name='create_session'),
     url(r'^welcome/(?P<session_id>[0-9]+)/$', views.welcome, name='welcome'),
     url(r'^edit/(?P<session_id>[0-9]+)/$', views.edit, name='edit'),
